+++ src/modules/mime/mimetic/codec/qp.h
@@ -360,7 +360,7 @@ class Decoder: public buffered_codec, public chainable_codec<Encoder>
                     m_nl = c;
                     return;
                 } else {
-                    int len = m_prev.length();
+                    size_t len = m_prev.length();
                     if(!len || m_prev[0] != '=')
                         hardLineBrk(out);
                     m_prev.clear();
@@ -468,7 +468,7 @@ public:
               malformed, just print '=' and ignore trailing
               blanks (rfc2045 6.7 (3) )
         */
-        int len = m_prev.length();
+        size_t len = m_prev.length();
         if(len)
         {
             if(len == 1)
