+++ src/modules/mime/mimetic/strutils.h
@@ -135,7 +135,7 @@ inline std::string parenthed(const std::string& s)
 /// removes double quotes
 inline std::string remove_dquote(const std::string& s)
 {
-    int len = s.length();
+    size_t len = s.length();
     if( len < 2)
         return s;
     if(s[0] == '"' && s[len-1] == '"')
@@ -155,7 +155,7 @@ inline std::string remove_external_blanks(const std::string& in)
     if(!in.length())
         return in;
     std::string s = in;
-    int beg = 0, end = s.length();
+    size_t beg = 0, end = s.length();
     for(; beg < end; ++beg)
         if(s[beg] != ' ' && s[beg] != '\t')
             break;
